split up .env variables in config and secrets
Before, configuration env variables were only present in the example .env file. You were supposed to create a separate .env file with the actual values. This meant that, when updating the example during development, an extra step is needed to update the local .env file to include the new changes during a deploy (e.g. a new plugin in the plugin list).

To avoid this extra step, the config and secret variables are split up: config contains what can be committed. If changed, a pull on a running instance will automatically have the new values, which are directly related to how the app should be run.

.env.secrets contains the secrets that should be set per running instance and will not update as often.