use .env instead of .env.config
Some vars are used in docker-compose.yml directly, so have to be part of the environment, meaning a .env file has to be used explicitely.

Alternatively this could be exchanged by setting the versions explicitely in the docker-compose file. But this keeps the setup closer to what it was before.